Refactoring: common ChainId value type.

+++ src/sentry/sentry_client_impl.rs
@@ -28,7 +28,7 @@ impl SentryClient for SentryClientImpl {
         };
 
         let status_data = grpc_sentry::StatusData {
-            network_id: u64::from(status.chain_fork_config.id.0),
+            network_id: status.chain_fork_config.id.0,
             total_difficulty: Some(grpc_types::H256::from(status.total_difficulty)),
             best_hash: Some(grpc_types::H256::from(status.best_hash)),
             fork_data: Some(fork_data),
